refactor functional tests - break into sep files
