+++ test/functional/cleanup.js
@@ -1,0 +1,8 @@
+import test from 'tape';
+
+export default function cleanup() {
+  test('force kill amplify process', (t) => {
+    t.end();
+    process.exit(0);
+  });
+}
