+++ test/functional/getEndpointPath.js
@@ -1,0 +1,3 @@
+const getEndpointPath = rec => `/foto/${rec.username}/${rec.id}`;
+
+export default getEndpointPath;
