+++ test/functional/people.js
@@ -1,0 +1,83 @@
+import test from 'tape';
+import formatError from './formatError';
+
+export default function peopleTests(setupData, api) {
+  let people;
+
+  test('getPeople', (t) => {
+    api.get(setupData.apiUrl, '/people')
+      .then((responseBody) => {
+        people = responseBody;
+        t.ok(Array.isArray(responseBody), 'people array');
+        t.end();
+      })
+      .catch(formatError);
+  });
+
+  // These test are conditional on people length
+  // this is a temp fix for occasional race condition - eg: https://travis-ci.org/mbudm/fotopia-serverless/jobs/426215588
+  // sometimes the faces lambda - that creates the people object in s3 is not complete
+  // before the functional tests get to this point. Until I think of a more robust option,
+  // cordoning off these two tests
+  const updatedPerson = {
+    name: 'Jacinta Dias',
+  };
+
+  test('updatePerson', (t) => {
+    if (people.length > 0) {
+      api.put(setupData.apiUrl, `/person/${people[0].id}`, { body: updatedPerson })
+        .then((responseBody) => {
+          t.ok(responseBody, 'update person ok');
+          t.end();
+        })
+        .catch(formatError);
+    } else {
+      t.end();
+    }
+  });
+
+  test('getPeople - check updated name', (t) => {
+    if (people.length > 0) {
+      api.get(setupData.apiUrl, '/people')
+        .then((responseBody) => {
+          people = responseBody;
+          const personInResponse = responseBody.find(person => person.id === people[0].id);
+          t.equal(personInResponse.name, updatedPerson.name, 'updated name');
+          t.end();
+        })
+        .catch(formatError);
+    } else {
+      t.end();
+    }
+  });
+
+  test('peopleMerge', (t) => {
+    if (people.length > 1) {
+      const body = people
+        .reduce((accum, person) => (accum.length < 2 ? accum.concat(person.id) : accum), []);
+      api.post(setupData.apiUrl, '/people/merge', {
+        body,
+      })
+        .then((responseBody) => {
+          t.ok(responseBody, 'peopleMerge person ok');
+          t.end();
+        })
+        .catch(formatError);
+    } else {
+      t.end();
+    }
+  });
+
+  test('getPeople - check peopleMerge', (t) => {
+    if (people.length > 1) {
+      api.get(setupData.apiUrl, '/people')
+        .then((responseBody) => {
+          t.equal(responseBody.length, people.length - 1, 'one less person');
+          t.end();
+        })
+        .catch(formatError);
+    } else {
+      t.end();
+    }
+  });
+}
